# server/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bloom Smart Habit REST API",
    "version": "1.3.0",
    "description": "Versioned API contract for web and native mobile clients."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "Authenticated user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "operationId": "updateProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProfileUpdateInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated." },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/profile/avatar": {
      "get": {
        "operationId": "getProfileAvatar",
        "responses": {
          "200": {
            "description": "Private normalized profile photo.",
            "content": {
              "image/webp": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "put": {
        "operationId": "uploadProfileAvatar",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile photo updated." },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "operationId": "deleteProfileAvatar",
        "responses": {
          "200": { "description": "Profile photo removed." },
          "401": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OpaqueSessionToken"
      }
    },
...
    "responses": {
      "Error": {
        "description": "Consistent error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ProfileUpdateInput": {
        "type": "object",
        "required": ["name", "timezone", "units"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "timezone": { "type": "string", "minLength": 1, "maxLength": 100 },
          "units": { "type": "string", "enum": ["metric", "imperial"] }
        }
      },
...
      "Profile": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "timezone",
          "units",
          "account_created_at",
          "has_avatar"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "units": {
            "type": "string",
            "enum": ["metric", "imperial"]
          },
          "account_created_at": {
            "type": "string",
            "format": "date-time"
          },
          "has_avatar": {
            "type": "boolean"
          },
          "avatar_updated_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "suspended_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "onboarding_completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "goal_preferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalPreference"
            }
          },
          "starting_pace": {
            "$ref": "#/components/schemas/StartingPace"
          },
          "religion_preference": {
            "$ref": "#/components/schemas/ReligionPreference"
          },
          "daily_digest_time": {
            "type": "string"
          },
          "daily_digest_enabled": {
            "type": "boolean"
          },
          "latitude": {
            "type": [
              "number",
              "string",
              "null"
            ]
          },
          "longitude": {
            "type": [
              "number",
              "string",
              "null"
            ]
          },
          "madhab": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Madhab"
              },
              {
                "type": "null"
              }
            ]
          },
          "prayer_calculation_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrayerCalculationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "prayer_reminders": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "prayer_name",
                "enabled",
                "offset_minutes"
              ],
              "properties": {
                "prayer_name": {
                  "$ref": "#/components/schemas/PrayerName"
                },
                "enabled": {
                  "type": "boolean"
                },
                "offset_minutes": {
                  "type": "integer"
                }
              }
            }
          },
          "push_enabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": true
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "correlationId",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "GoalPreference": {
        "type": "string",
        "enum": [
          "movement",
          "nutrition",
          "learning",
          "sleep",
          "mindfulness"
        ]
      },
      "StartingPace": {
        "type": "string",
        "enum": [
          "light",
          "balanced",
          "ambitious"
        ]
      },
      "ReligionPreference": {
        "type": "string",
        "enum": [
          "muslim",
          "other",
          "unspecified"
        ]
      },
      "Madhab": {
        "type": "string",
        "enum": [
          "hanafi",
          "shafi",
          "maliki",
          "hanbali"
        ]
      },
      "PrayerCalculationMethod": {
        "type": "string",
        "enum": [
          "karachi",
          "muslim_world_league",
          "egyptian",
          "umm_al_qura",
          "dubai",
          "qatar",
          "kuwait",
          "moonsighting_committee",
          "singapore",
          "turkey",
          "tehran",
          "north_america"
        ]
      },
      "PrayerName": {
        "type": "string",
        "enum": [
          "fajr",
          "dhuhr",
          "asr",
          "maghrib",
          "isha"
        ]
      },
...
    }
  }
}